third_party/blink/web_tests/fast/dom/Window/window-postmessage-clone-expected.txt

Tests that we clone object hierarchies
PASS: 'postMessage((function(){}))' threw DataCloneError: Failed to execute 'postMessage' on 'Window': function(){} could not be cloned.
PASS: 'postMessage(arrayBuffer)' threw DataCloneError: Failed to execute 'postMessage' on 'Window': An ArrayBuffer is detached and could not be cloned.
PASS: 'postMessage("data")' threw DataCloneError: Failed to execute 'postMessage' on 'Window': ArrayBuffer at index 1 is a duplicate of an earlier ArrayBuffer.
PASS: 'postMessage(uint8Array)' threw DataCloneError: Failed to execute 'postMessage' on 'Window': An ArrayBuffer is detached and could not be cloned.
PASS: 'postMessage(window)' threw DataCloneError: Failed to execute 'postMessage' on 'Window': #<Window> could not be cloned.
PASS: 'postMessage(({get a() { throw "x" }}))' threw x
PASS: 'postMessage((function() {return {get a() { throw "accessor-exn"; }};})())' threw accessor-exn
PASS: 'postMessage((function() {var obja = {get p() { throw "accessor-exn"; }}; return {get a() { return obja; }};})())' threw accessor-exn
PASS: lastIndex initially correct: was 3
PASS: sanity on file reference equality
PASS: eventData is null of type object
PASS: eventData is undefined of type undefined
PASS: eventData is 1 of type number
PASS: eventData is true of type boolean
PASS: eventData is 1 of type string
PASS: eventData is [object Object] of type object
PASS: eventData is [object Object] of type object
PASS: eventData is [object Object] of type object
PASS: eventData is [object Object] of type object
PASS: eventData is [object Object] of type object
PASS: eventData is [object Object] of type object
PASS: eventData is [object Object] of type object
PASS: eventData is [object Object] of type object
PASS: eventData is [object Object] of type object
PASS: eventData is of type object
PASS: eventData is a,a,b,a,b of type object
PASS: eventData is a,a,b,[object Object] of type object
PASS: eventData is 1,2,3 of type object
PASS: eventData is ,,1 of type object
PASS: eventData is ReferenceError: badref is not defined of type object
PASS: eventData is data of type string
PASS: eventData is data of type string
PASS: eventData is 2009-02-13T23:31:30.000Z of type object
PASS: eventData is [object Object] of type object
PASS: eventData is true of type object
PASS: eventData is false of type object
PASS: eventData is gnirts of type object
PASS: eventData is 42 of type object
PASS: eventData is === to eventData.self
PASS: eventData is === to eventData[0]
PASS: eventData is === to eventData.self
PASS: eventData.graph1 is === to eventData.graph2
PASS: eventData[0] is === to eventData[1]
PASS: eventData is [object ImageData] of type object
PASS: eventData is 0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30,31,32,33,34,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,58,59,60,61,62,63,64,65,66,67,68,69,70,71,72,73,74,75,76,77,78,79,80,81,82,83,84,85,86,87,88,89,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,105,106,107,108,109,110,111,112,113,114,115,116,117,118,119,120,121,122,123,124,125,126,127,128,129,130,131,132,133,134,135,136,137,138,139,140,141,142,143,144,145,146,147,148,149,150,151,152,153,154,155,156,157,158,159,160,161,162,163,164,165,166,167,168,169,170,171,172,173,174,175,176,177,178,179,180,181,182,183,184,185,186,187,188,189,190,191,192,193,194,195,196,197,198,199,200,201,202,203,204,205,206,207,208,209,210,211,212,213,214,215,216,217,218,219,220,221,222,223,224,225,226,227,228,229,230,231,232,233,234,235,236,237,238,239,240,241,242,243,244,245,246,247,248,249,250,251,252,253,254,255,0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30,31,32,33,34,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,58,59,60,61,62,63,64,65,66,67,68,69,70,71,72,73,74,75,76,77,78,79,80,81,82,83,84,85,86,87,88,89,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,105,106,107,108,109,110,111,112,113,114,115,116,117,118,119,120,121,122,123,124,125,126,127,128,129,130,131,132,133,134,135,136,137,138,139,140,141,142,143 of type object
PASS: eventData is [object ImageData] of type object
PASS: eventData is 0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30,31,32,33,34,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,58,59,60,61,62,63,64,65,66,67,68,69,70,71,72,73,74,75,76,77,78,79,80,81,82,83,84,85,86,87,88,89,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,105,106,107,108,109,110,111,112,113,114,115,116,117,118,119,120,121,122,123,124,125,126,127,128,129,130,131,132,133,134,135,136,137,138,139,140,141,142,143,144,145,146,147,148,149,150,151,152,153,154,155,156,157,158,159,160,161,162,163,164,165,166,167,168,169,170,171,172,173,174,175,176,177,178,179,180,181,182,183,184,185,186,187,188,189,190,191,192,193,194,195,196,197,198,199,200,201,202,203,204,205,206,207,208,209,210,211,212,213,214,215,216,217,218,219,220,221,222,223,224,225,226,227,228,229,230,231,232,233,234,235,236,237,238,239,240,241,242,243,244,245,246,247,248,249,250,251,252,253,254,255,0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,25,26,27,28,29,30,31,32,33,34,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,58,59,60,61,62,63,64,65,66,67,68,69,70,71,72,73,74,75,76,77,78,79,80,81,82,83,84,85,86,87,88,89,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,105,106,107,108,109,110,111,112,113,114,115,116,117,118,119,120,121,122,123,124,125,126,127,128,129,130,131,132,133,134,135,136,137,138,139,140,141,142,143 of type object
PASS: Cloned Blob size equal to the original size.
PASS: Cloned File size equal to the original size.
PASS: eventData is 42 of type number
PASS: eventData is 42 of type number
PASS: evalThunk OK
PASS: eventData is [object Object] of type object
PASS: length correct
PASS: evaluation order OK
PASS: evaluation order OK/accessor reached
PASS: evaluation order OK
PASS: length correct
PASS: index 0 OK
PASS: accessor reached
PASS: index 2 undefined
PASS: length correct
PASS: index 0 OK
PASS: index 1 OK
PASS: index 2 OK
PASS: basic properties OK
PASS: one accessor was run
PASS: property was removed
PASS: property was removed
PASS: accessor was run
PASS: dynamic property not sent
PASS: reference equality preserved
PASS: accessors used
PASS: reference equality preserved (opposite order)
PASS: accessors used (opposite order)
PASS: accessor value was not nullified
PASS: non-accessor value was not nullified
PASS: accessor value was not nullified
PASS: non-accessor value was not nullified
PASS: non-accessor value was not nullified
PASS: accessor value was not nullified
PASS: non-accessor value was not nullified
PASS: accessor value was not nullified
PASS: accessor value was not nullified
PASS: accessor value was not nullified
PASS: non-accessor value was not nullified
PASS: non-accessor value was not nullified
PASS: non-accessor value is defined
PASS: non-accessor value was not nullified
PASS: non-accessor value is defined
PASS: accessor value was not nullified
PASS: non-accessor value was not nullified
PASS: accessor value was not nullified
PASS: window.pcalled === true
PASS: window.acalled === true
PASS: window.bcalled === true
PASS: Boolean values correct (0)
PASS: Boolean values correct (3)
PASS: Boolean values correct (1,2)
PASS: Boolean values correct (4,5)
PASS: References to Booleans correct (0,1)
PASS: References to Booleans correct (3,4)
PASS: References to Booleans correct (0,2)
PASS: References to Booleans correct (3,5)
PASS: Number values correct (0)
PASS: Number values correct (0,2)
PASS: References to numbers correct (0,1)
PASS: References to numbers correct (0,2)
PASS: String values correct (0)
PASS: String values correct (0,2)
PASS: References to strings correct (0,1)
PASS: References to strings correct (0,2)
PASS: String object properties not serialized
PASS: Number object properties not serialized
PASS: Boolean object properties not serialized
PASS: Date values correct (0)
PASS: Date values correct (1)
PASS: References to dates correct (0)
PASS: References to dates correct (1)
PASS: Regexp value correct (0)
PASS: References to regexps correct (0)
PASS: References to regexps correct (1)
PASS: global set (0)
PASS: global set (1)
PASS: ignoreCase set (0)
PASS: ignoreCase set (1)
PASS: multiline set (0)
PASS: multiline set (1)
PASS: lastIndex correct (0)
PASS: lastIndex correct (1)
PASS: eventData is 2 of type number
PASS: file references transfer
PASS: FileList should not respect reference equality
PASS: FileList should not respect reference equality
PASS: FileList respects self-reference equality
PASS: FileList length sent correctly
PASS: FileList should not respect reference equality (0)
PASS: FileList should not respect reference equality (1)
PASS: FileList preserves order and data (name0)
PASS: FileList preserves order and data (name1)
PASS: FileList preserves order and data (date0)
PASS: FileList preserves order and data (date1)
PASS: eventData is done of type string
